HttpClient: introduce abstract base class
All implementations of HttpClient should inherit from a common base
class to enforce interface contracts required by Consul implementations
as well as provide common functionality.

This commit introduces a Python 2/3 compatible base class implementation
and replaces all current implementations to inherit from
consul.base.HttpClient.
